refactor server code: create router dir
