+++ server/routers/__init__.py
@@ -1,0 +1,2 @@
+from .absa_pipeline import router as absa_pipeline
+from .gpt_absa import router as gpt_absa
