+++ server/utils/__init__.py
@@ -3,3 +3,4 @@ from .reviews_scrapper import review_scrapper
 from .text_cleaner import text_process
 from .text_segmentation import segment_review
 from .word_cloud import get_word_cloud
+from .pipeline import pipeline
